pyproject.toml
[tool.poetry]
name = "gpt-copywriter"
version = "1"
description = ""
authors = ["Michael Miller"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.12.0"
pymupdf = "^1.23.21"
python-dotenv = "^1.0.1"
flask = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
